+++ HarmonEye/src/test/java/com/harmoneye/math/filter/DownsamplerTest.java
@@ -8,18 +8,18 @@ public class DownsamplerTest {
 
 	@Test
 	public void downsampleSomeSignal() {
-		float freq = 20.0f;
-		float sampleFreq = 100.0f;
-		float duration = 2.0f;
-		float[] signal = generateSinWave(freq, sampleFreq, duration);
-		float[] expectedSignal = generateSinWave(freq, 0.5f * sampleFreq, duration);
+		double freq = 20.0f;
+		double sampleFreq = 100.0f;
+		double duration = 2.0f;
+		double[] signal = generateSinWave(freq, sampleFreq, duration);
+		double[] expectedSignal = generateSinWave(freq, 0.5f * sampleFreq, duration);
 
 		System.out.println("Original signal:");
 		System.out.println("RMS: " + computeRms(signal));
 		printSignal(signal);
 
 		Decimator downsampler = Decimator.withDefaultFilter();
-		float[] downsampledSignal = downsampler.decimate(signal, null);
+		double[] downsampledSignal = downsampler.decimate(signal, null);
 
 		System.out.println("Decimated signal:");
 		System.out.println("RMS: " + computeRms(downsampledSignal));
