+++ HarmonEye/src/test/java/com/harmoneye/math/filter/ZeroPhaseButterworthFilterTest.java
@@ -10,8 +10,8 @@ public class ZeroPhaseButterworthFilterTest {
 		Filter filter = new ZeroPhaseFilter(new ButterworthFilter());
 
 		for (int i = 0; i < 20; i++) {
-			float frequency = i * 0.5f;
-			float[] signal = generateSinWave(frequency, 20.0f, 2.0f);
+			double frequency = i * 0.5f;
+			double[] signal = generateSinWave(frequency, 20.0f, 2.0f);
 
 			filter.filter(signal, signal);
 
@@ -24,18 +24,18 @@ public class ZeroPhaseButterworthFilterTest {
 		Filter butterworthFilter = new ButterworthFilter();
 		Filter zeroPhasefilter = new ZeroPhaseFilter(butterworthFilter);
 
-		float[] signal = new float[20];
+		double[] signal = new double[20];
 		signal[10] = 1.0f;
 
 		System.out.println("Original signal - spike:");
 		printSignal(signal);
 		
-		float[] bwFiltered = butterworthFilter.filter(signal, null);
+		double[] bwFiltered = butterworthFilter.filter(signal, null);
 
 		System.out.println("Single-pass Butterworth:");
 		printSignal(bwFiltered);
 
-		float[] zpFiltered = zeroPhasefilter.filter(signal, null);
+		double[] zpFiltered = zeroPhasefilter.filter(signal, null);
 
 		System.out.println("Forward-and-reverse two-pass Butterworth:");
 		printSignal(zpFiltered);
@@ -47,22 +47,22 @@ public class ZeroPhaseButterworthFilterTest {
 	 * @param length in seconds
 	 * @return
 	 */
-	private float[] generateSinWave(float frequency, float samplingFreq,
-		float length) {
+	private double[] generateSinWave(double frequency, double samplingFreq,
+		double length) {
 		int sampleCount = (int) Math.ceil(length * samplingFreq);
-		float[] signal = new float[sampleCount];
+		double[] signal = new double[sampleCount];
 		for (int i = 0; i < signal.length; i++) {
-			signal[i] = (float) Math.sin(i * 2 * Math.PI * frequency / samplingFreq);
+			signal[i] = (double) Math.sin(i * 2 * Math.PI * frequency / samplingFreq);
 		}
 		return signal;
 	}
 
-	private float computeRms(float[] samples) {
-		float sum = 0;
-		for (float amplitude : samples) {
+	private double computeRms(double[] samples) {
+		double sum = 0;
+		for (double amplitude : samples) {
 			sum += amplitude * amplitude;
 		}
-		return (float) Math.sqrt(sum / (float) samples.length);
+		return (double) Math.sqrt(sum / (double) samples.length);
 	}
 
 }
